+++ src/lib.rs
@@ -311,6 +311,36 @@ macro_rules! impl_erased_set {
                 unsafe { &*ptr }
             }
 
+            /// Inserts a value computed from `f` into the set if it does not contain
+            /// a value of type `T`, then returns a reference to the value in the set.
+            ///
+            /// # Examples
+            ///
+            /// ```
+            #[doc = concat!("use ", module_path!(), "::", stringify!($name), ";")]
+            ///
+            #[doc = concat!("let mut erased_set = ", stringify!($name), "::new();")]
+            /// assert_eq!(erased_set.get_or_insert_with(|| String::from("abc")), &"abc");
+            /// assert_eq!(erased_set.get_or_insert_with(|| String::from("def")), &"abc");
+            /// ```
+            #[must_use]
+            pub fn get_or_insert_with<T>(&mut self, f: impl FnOnce() -> T) -> &T
+            where
+                T: Any $(+ $bounds)*,
+            {
+                let boxed_any: &Box<dyn Any $(+ $bounds)*> = self
+                    .0
+                    .entry(TypeId::of::<T>())
+                    .or_insert_with(|| Box::new(f()));
+
+                // Sanity check
+                debug_assert!(boxed_any.as_ref().is::<T>());
+
+                let ptr = (boxed_any.as_ref() as *const dyn Any).cast::<T>();
+
+                unsafe { &*ptr }
+            }
+
             /// Returns a mutable reference to an instance of `T`.
             ///
             /// If the set does not have an instance of `T`, [`None`] is returned.