+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `ErasedSet::shrink_to`
 - `ErasedSet::get_or_insert`
+- `ErasedSet::get_or_insert_with`
 
 ## [0.6.0] - 2022-05-11